requirements.txt
selenium==4.4.2
pytest
pytest-xdist
allure-pytest
webdriver-manager